+++ src/equality.c
@@ -1,5 +1,5 @@
 #include "equality.h"
 
-bool FloatEquality(float f1, float f2) {
-    return fabsf(f1 - f2) < EQUALITY_EPSILON;
+bool FloatEquality(double f1, double f2) {
+    return fabs(f1 - f2) < EQUALITY_EPSILON;
 }
